refac: Add findings file and remove redundant imports
